SwaggerUI BugFix URL Creation
In the used Swagger Version the URL in the Curl commands contains the "/mbp" twice, so all created curl commands are not working correctly.
This is beacuse the "/mbp" is added with the context Path and also the base Path.

In this commit the part of the url is removed and the UI is working again.